+++ tests/testthat/test-provider-azure.R
@@ -1,17 +1,55 @@
-test_that("can make simple request", {
-  chat <- chat_azure(
-    system_prompt = "Be as terse as possible; no punctuation",
-    endpoint = "https://ai-hwickhamai260967855527.openai.azure.com",
-    deployment_id = "gpt-4o-mini"
-  )
-  resp <- chat$chat("What is 1 + 1?")
-  expect_match(resp, "2")
-  expect_equal(chat$last_turn()@tokens, c(27, 1))
+# Getting started --------------------------------------------------------
 
-  resp <- sync(chat$chat_async("What is 1 + 1?"))
+test_that("can make simple request", {
+  chat <- chat_azure_test("Be as terse as possible; no punctuation")
+  resp <- chat$chat("What is 1 + 1?", echo = FALSE)
   expect_match(resp, "2")
-  expect_equal(chat$last_turn()@tokens, c(44, 1))
+  expect_true(all(chat$last_turn()@tokens >= 1))
 })
+
+test_that("can make simple streaming request", {
+  chat <- chat_azure_test("Be as terse as possible; no punctuation")
+  resp <- coro::collect(chat$stream("What is 1 + 1?"))
+  expect_match(paste0(unlist(resp), collapse = ""), "2")
+})
+
+# Common provider interface -----------------------------------------------
+
+test_that("defaults are reported", {
+  expect_snapshot(. <- chat_azure_test())
+})
+
+test_that("respects turns interface", {
+  chat_fun <- chat_azure_test
+
+  test_turns_system(chat_fun)
+  test_turns_existing(chat_fun)
+})
+
+test_that("all tool variations work", {
+  chat_fun <- chat_azure_test
+
+  test_tools_simple(chat_fun)
+  test_tools_async(chat_fun)
+  test_tools_parallel(chat_fun)
+  test_tools_sequential(chat_fun, total_calls = 6)
+})
+
+test_that("can extract data", {
+  chat_fun <- chat_azure_test
+
+  test_data_extraction(chat_fun)
+})
+
+test_that("can use images", {
+  skip("Run manually; 24 hour rate limit")
+  chat_fun <- chat_azure_test
+
+  httr2::with_verbosity(test_images_inline(chat_fun), 2)
+  test_images_remote(chat_fun)
+})
+
+# Authentication --------------------------------------------------------------
 
 test_that("Azure request headers are generated correctly", {
   turn <- Turn(